add docs for Relation.incremental, remove redundant docstrings, consolidate the tests
