+++ tests/load/test_read_interfaces.py
@@ -21,7 +21,6 @@ from dlt.common.schema.typing import TTableFormat
 
 from dlt.common.utils import uniq_id
 from dlt.extract.incremental import Incremental
-from dlt.extract.incremental.sql import to_sqlglot_filter
 from dlt.extract.source import DltSource
 from dlt.dataset.exceptions import LineageFailedException
 
@@ -848,96 +847,8 @@ def test_where(populated_pipeline: Pipeline) -> None:
 
 @pytest.mark.no_load
 @pytest.mark.essential
-def test_to_sqlglot_filter_on_dataset(populated_pipeline: Pipeline) -> None:
-    """End-to-end: to_sqlglot_filter(incr) applied via dataset().items.where(expr)."""
-    _run_filter_assertions(populated_pipeline)
-
-
-def _run_filter_assertions(populated_pipeline: Pipeline) -> None:
-    items = populated_pipeline.dataset().items
-    caps = populated_pipeline.dataset().sql_client.capabilities
-    total_records = _total_records(populated_pipeline.destination.destination_type)
-    last_dt = ITEMS_EPOCH + timedelta(seconds=total_records - 1)
-
-    # post-run state matches what `bind()` persisted: start_value snapshot == last_value
-    cached_state: IncrementalColumnState = {
-        "initial_value": ITEMS_EPOCH,
-        "last_value": last_dt,
-        "start_value": last_dt,
-        "unique_hashes": [],
-    }
-
-    def _bind(incr: Incremental[Any], instance_start_value: Any = None) -> Incremental[Any]:
-        # mirror Incremental.bind(): instance start_value is the lag-applied lower
-        # (or raw last_value when no lag), cached_state["start_value"] is the raw snapshot
-        incr._cached_state = copy(cached_state)
-        incr.start_value = instance_start_value if instance_start_value is not None else last_dt
-        return incr
-
-    # bound, no lag: lower == upper == last_dt -> "no new data" filter, 0 rows
-    incr = _bind(dlt.sources.incremental[pendulum.DateTime]("created_at"))
-    expr = to_sqlglot_filter(incr, destination_capabilities=caps)
-    assert expr is not None
-    assert items.where(expr).fetchall() == []
-
-    # 1. lag — 5 seconds backward; bind() would set start_value = last_dt - 5s
-    lagged_start = last_dt - timedelta(seconds=5)
-    incr_lag = _bind(
-        dlt.sources.incremental[pendulum.DateTime]("created_at", lag=5.0),
-        instance_start_value=lagged_start,
-    )
-    expr = to_sqlglot_filter(incr_lag, destination_capabilities=caps)
-    # filter: created_at >= last_dt-5s AND < last_dt -> positions total-6..total-2 = 5 rows
-    assert len(items.where(expr).fetchall()) == 5
-    # apply_lag=False reads raw cached state["start_value"] (= last_dt) -> 0 rows
-    expr_raw = to_sqlglot_filter(incr_lag, apply_lag=False, destination_capabilities=caps)
-    assert items.where(expr_raw).fetchall() == []
-
-    # 2. no upper bound — fresh, unbound, just initial_value
-    # filter: created_at >= ITEMS_EPOCH -> all rows
-    incr_unbound = dlt.sources.incremental[pendulum.DateTime](
-        "created_at", initial_value=ITEMS_EPOCH
-    )
-    expr = to_sqlglot_filter(incr_unbound, destination_capabilities=caps)
-    assert len(items.where(expr).fetchall()) == total_records
-
-    # 3. open start, closed end — fresh, unbound, range modifiers on both ends
-    range_start_dt = ITEMS_EPOCH + timedelta(seconds=10)
-    range_end_dt = ITEMS_EPOCH + timedelta(seconds=20)
-    incr_range = dlt.sources.incremental[pendulum.DateTime](
-        "created_at",
-        initial_value=range_start_dt,
-        end_value=range_end_dt,
-        range_start="open",
-        range_end="closed",
-    )
-    expr = to_sqlglot_filter(incr_range, destination_capabilities=caps)
-    # filter: created_at > t+10s AND created_at <= t+20s -> positions 11..20 = 10 rows
-    arrow_tbl = items.where(expr).select("created_at").order_by("created_at").arrow()
-    actual_dts = arrow_tbl["created_at"].to_pylist()
-    expected_dts = [ITEMS_EPOCH + timedelta(seconds=i) for i in range(11, 21)]
-
-    # normalize to Unix int_timestamp — destinations differ in how they return tz
-    # information (naive vs aware, timezone class), so compare instants
-    def _ts(d: Any) -> int:
-        if isinstance(d, str):
-            d = pendulum.parse(d)
-        if d.tzinfo is None:
-            d = d.replace(tzinfo=pendulum.UTC)
-        return int(d.timestamp())
-
-    assert [_ts(d) for d in actual_dts] == [_ts(d) for d in expected_dts]
-
-
-@pytest.mark.no_load
-@pytest.mark.essential
 def test_relation_incremental_datetime_on_dataset(populated_pipeline: Pipeline) -> None:
-    """End-to-end: dataset.table('items').incremental(<datetime cursor>) on every destination.
-
-    Mirrors `test_to_sqlglot_filter_on_dataset` but exercises the `_build_incremental_condition`
-    path.
-    """
-    # TODO: consider merging with `test_to_sqlglot_filter_on_dataset`.
+    """End-to-end: dataset.table('items').incremental(<datetime cursor>) on every destination."""
     items = populated_pipeline.dataset().items
     total_records = _total_records(populated_pipeline.destination.destination_type)
     last_dt = ITEMS_EPOCH + timedelta(seconds=total_records - 1)
@@ -955,13 +866,11 @@ def test_relation_incremental_datetime_on_dataset(populated_pipeline: Pipeline) 
         incr.start_value = instance_start_value if instance_start_value is not None else last_dt
         return incr
 
-    # 1. bound, no lag, no end_value — lower = last_dt, no upper
-    #    diverges from to_sqlglot_filter (0 rows): dataset path keeps the last row
+    # 1. bound, no lag, no end_value — lower = last_dt, no upper, keeps the last row
     incr = _bind(dlt.sources.incremental[pendulum.DateTime]("created_at"))
     assert len(items.incremental(incr).fetchall()) == 1
 
-    # 2. lag — start_value = last_dt - 5s; no upper bound on dataset side
-    #    diverges from to_sqlglot_filter (5 rows): includes last_dt itself
+    # 2. lag — start_value = last_dt - 5s; no upper bound; includes last_dt itself
     lagged_start = last_dt - timedelta(seconds=5)
     incr_lag = _bind(
         dlt.sources.incremental[pendulum.DateTime]("created_at", lag=5.0),
@@ -969,7 +878,7 @@ def test_relation_incremental_datetime_on_dataset(populated_pipeline: Pipeline) 
     )
     assert len(items.incremental(incr_lag).fetchall()) == 6
 
-    # 3. unbound, initial_value only — pre-fix this raised IncrementalUnboundError
+    # 3. unbound, initial_value only
     incr_unbound = dlt.sources.incremental[pendulum.DateTime](
         "created_at", initial_value=ITEMS_EPOCH
     )
